+++ package.json
@@ -139,7 +139,7 @@
     "@patternfly/react-table": "^4.16.20",
     "@redhat-cloud-services/entitlements-client": "^1.0.75",
     "@redhat-cloud-services/frontend-components": "^2.4.12",
-    "@redhat-cloud-services/frontend-components-inventory": "^2.4.16",
+    "@redhat-cloud-services/frontend-components-inventory": "^2.4.17",
     "@redhat-cloud-services/frontend-components-inventory-compliance": "^2.2.8",
     "@redhat-cloud-services/frontend-components-inventory-general-info": "^2.2.2",
     "@redhat-cloud-services/frontend-components-inventory-insights": "^2.6.4",
@@ -147,7 +147,7 @@
     "@redhat-cloud-services/frontend-components-inventory-vulnerabilities": "^1.47.0",
     "@redhat-cloud-services/frontend-components-notifications": "^2.2.2",
     "@redhat-cloud-services/frontend-components-remediations": "^2.2.3",
-    "@redhat-cloud-services/frontend-components-utilities": "^2.2.5",
+    "@redhat-cloud-services/frontend-components-utilities": "^2.2.6",
     "@redhat-cloud-services/host-inventory-client": "^1.0.79",
     "@redhat-cloud-services/keycloak-js": "0.0.3",
     "@redhat-cloud-services/rbac-client": "^1.0.75",
