+++ src/js/App/GlobalFilter/tagsApi.js
@@ -1,33 +1,19 @@
+/* eslint-disable camelcase */
 import instance from '@redhat-cloud-services/frontend-components-utilities/files/interceptors';
-import { INVENTORY_API_BASE } from './constants';
+import { flatTags, INVENTORY_API_BASE } from './constants';
 import { TagsApi, SapSystemApi } from '@redhat-cloud-services/host-inventory-client';
+import { generateFilter } from '@redhat-cloud-services/frontend-components-utilities/files/helpers';
 export const tags = new TagsApi(undefined, INVENTORY_API_BASE, instance);
 export const sap = new SapSystemApi(undefined, INVENTORY_API_BASE, instance);
-import flatMap from 'lodash/flatMap';
-
-const flatTags = (filter, encode = false, format = false) => {
-    const { Workloads, SID, ...tags } = filter;
-    const mappedTags = flatMap(
-        Object.entries({ ...tags, ...!format && { Workloads, SID } } || {}),
-        ([namespace, item]) => Object.entries(item)
-        .filter(([, { isSelected }]) => isSelected)
-        .map(([groupKey, { item, value: tagValue }]) => `${
-                namespace ? `${encode ? encodeURIComponent(namespace) : namespace}/` : ''
-            }${
-                encode ? encodeURIComponent(groupKey) : groupKey
-            }${
-                (item?.tagValue || tagValue) ? `=${encode ? encodeURIComponent(item?.tagValue || tagValue) : item?.tagValue || tagValue}` : ''
-            }`)
-    );
-    return format ? [
-        Workloads,
-        SID,
-        mappedTags
-    ] : mappedTags;
-};
 
 export function getAllTags({ search, activeTags, registeredWith } = {}, pagination = {}) {
     const [workloads, SID, selectedTags] = flatTags(activeTags, false, true);
+    const filter = {
+        system_profile: {
+            ...workloads?.SAP?.isSelected && { sap_system: true },
+            sap_sids: SID
+        }
+    };
     return tags.apiTagGetTags(
         selectedTags, // tag filer
         'tag',
@@ -38,34 +24,52 @@ export function getAllTags({ search, activeTags, registeredWith } = {}, paginati
         search,
         registeredWith,
         undefined,
-        {}
+        {
+            query: generateFilter(filter)
+        }
     );
 }
 
 export function getAllSIDs({ activeTags, registeredWith } = {}, pagination = {}) {
     const [workloads, SID, selectedTags] = flatTags(activeTags, false, true);
-
+    const filter = {
+        system_profile: {
+            ...workloads?.SAP?.isSelected && { sap_system: true },
+            sap_sids: SID
+        }
+    };
     return sap.apiSystemProfileGetSapSids(
-        undefined, // tags
+        selectedTags, // tags
         (pagination && pagination.perPage) || 10,
         (pagination && pagination.page) || 1,
         undefined, // staleness,
         registeredWith,
         undefined,
-        {}
+        {
+            query: generateFilter(filter)
+        }
     );
 }
 
 export function getAllWorkloads({ activeTags, registeredWith } = {}, pagination = {}) {
     const [workloads, SID, selectedTags] = flatTags(activeTags, false, true);
 
+    const filter = {
+        system_profile: {
+            ...workloads?.SAP?.isSelected && { sap_system: true },
+            sap_sids: SID
+        }
+    };
+    console.log(filter, generateFilter(filter), 'ffffmmm');
     return sap.apiSystemProfileGetSapSystem(
-        undefined, // tags
+        selectedTags, // tags
         (pagination && pagination.perPage) || 10,
         (pagination && pagination.page) || 1,
         undefined, // staleness,
         registeredWith,
         undefined,
-        {}
+        {
+            query: generateFilter(filter)
+        }
     );
 }
